Full permission includes own permission

+++ src/Policies/Policy.php
@@ -34,11 +34,15 @@ class Policy
      */
     public function delete(User $user, $model)
     {
+        if ($user->hasPermissionTo('manage ' . static::$key) ) {
+            return true;
+        }
+
         if ($user->hasPermissionTo('manage own ' . static::$key)) {
             return $user->id == $model->user_id;
         }
 
-        return $user->hasPermissionTo('manage ' . static::$key);
+        return false;
     }
 
     /**
@@ -49,11 +53,15 @@ class Policy
      */
     public function forceDelete(User $user, $model)
     {
+        if ($user->hasPermissionTo('forceDelete ' . static::$key)) {
+            return true;
+        }
+
         if ($user->hasPermissionTo('manage own ' . static::$key) && $user->hasPermissionTo('forceDelete ' . static::$key)) {
             return $user->id == $model->user_id;
         }
 
-        return $user->hasPermissionTo('forceDelete ' . static::$key);
+        return false;
     }
 
     /**
@@ -64,11 +72,15 @@ class Policy
      */
     public function restore(User $user, $model)
     {
+        if ($user->hasPermissionTo('restore ' . static::$key)) {
+            return true;
+        }
+
         if ($user->hasPermissionTo('manage own ' . static::$key) && $user->hasPermissionTo('restore ' . static::$key)) {
             return $user->id == $model->user_id;
         }
 
-        return $user->hasPermissionTo('restore ' . static::$key);
+        return false;
     }
 
     /**
@@ -79,11 +91,15 @@ class Policy
      */
     public function update(User $user, $model)
     {
+        if ($user->hasPermissionTo('manage ' . static::$key)) {
+            return true;
+        }
+
         if ($user->hasPermissionTo('manage own ' . static::$key)) {
             return $user->id == $model->user_id;
         }
 
-        return $user->hasPermissionTo('manage ' . static::$key);
+        return false;
     }
 
     /**
@@ -94,11 +110,15 @@ class Policy
      */
     public function view(User $user, $model)
     {
+        if ($user->hasPermissionTo('view ' . static::$key)) {
+            return true;
+        }
+
         if ($user->hasPermissionTo('view own ' . static::$key)) {
             return $user->id == $model->user_id;
         }
 
-        return $user->hasPermissionTo('view ' . static::$key);
+        return false;
     }
 
     /**
